requirements.txt
langchain==0.2.10
langchain-core==0.2.23
langchain-community==0.2.10
langchain-text-splitters==0.2.0
langchain-groq==0.1.6
chromadb==0.5.4
sentence-transformers==3.0.1
python-dotenv==1.0.1
streamlit